styles for ethSpent table
