+++ src/components/EthSpentTable/EthSpentTable.module.scss
@@ -8,22 +8,49 @@
   width: 100%;
 }
 
-.transactionsTable:global(.ReactTable) {
+.ticker {
+  color: var(--waterloo);
+  margin-left: 6px;
+}
+
+.name {
+  color: var(--mirage);
+
+  &:hover {
+    color: var(--jungle-green);
+
+    & .ticker {
+      color: var(--jungle-green);
+    }
+  }
+}
+
+.ethSpentTable:global(.ReactTable) {
   border: none;
 
   :global {
     .rt-td {
       border: none !important;
       padding: 8px 0;
-      height: 39px;
       align-self: baseline;
       text-overflow: clip;
+      display: flex;
+      align-items: center;
 
       @include text('body-3');
 
       & > div {
         @include text('body-3');
       }
+
+      &:last-child {
+        flex-direction: column;
+        align-items: baseline;
+      }
+    }
+
+    .rt-tr-group {
+      max-height: none;
     }
 
     .rt-th {
@@ -35,12 +62,8 @@
       border: none;
       outline: none;
 
-      &:nth-child(4) {
-        padding-left: 7px;
-      }
-
-      &:nth-child(5) {
-        padding-left: 13px;
+      &:nth-child(6) {
+        padding-left: 11px;
       }
 
       @include text('caption');
@@ -59,7 +82,7 @@
         & > div {
           &::before,
           &::after {
-            content: "";
+            content: '';
             position: absolute;
             right: 0;
             border: 4px solid transparent;
