+++ src/pages/EthSpent.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import GetAssets from './../pages/assets/GetAssets'
-import EthSpentTable from './../components/EthSpentTable'
+import EthSpentTable from './../components/EthSpentTable/EthSpentTable'
 
 const EthSpent = () => (
   <GetAssets type='erc20' render={Assets => <EthSpentTable {...Assets} />} />
